Edges data

The edges dataset includes one observation per edge in the network of European Union legal documents. An edge is a connection between two documents, created by the newer of the two documents. The dataset tracks the following types of connections: Cites all or part of (relates to all types of documents), Changes text in clause of (relates to secondary laws), Replaces clause of (relates to seconary laws), Adds text to clause of (relates to secondary laws), Inserts new clause in (relates to secondary laws), Repeals all or part of (relates to secondary laws), Has legal basis in (relates to secondary laws), Interprets all or part of (relates to references for a preliminary ruling at the Court of Justice), Answers question referred by (relates to references fror a preliminary ruling at the Court of Justice), Determines compliance with (relates to actions for failure to fulfill obligations at the Court of Justice), Upholds all or part of (relates to court cases), Overturns all or part of (relates to court cases), and Transposes all or part of (relates to national implementing measures).

edges

Format

A data frame with 9 variables:

key_id

Numeric. An ID number that uniquely identifies each observation. Indicates the default sorting order.

outgoing_celex

String. The CELEX number of the outgoing document (the document that creates the edge).

outgoing_node_type_id

Numeric. An ID number that uniquely identifies the type of the outgoing node.

outgoing_node_type

String. The type of the outgoing node (the document that creates the edge). Possible values include: Treaty articles, Directives, Regulations. Decisions, Court of Justice judgments, Advocate general opinions, General Court judgments, National implementing measures, and National court decisions.

edge_type_id

Numeric. An ID number that uniquely identifies the type of the edge.

edge_type

String. The type of the edge. Possible values include: Cites all or part of, Changes text in clause of, Replaces clause of, Adds text to clause of, Inserts new clause in, Repeals all or part of, Has legal basis in, Interprets all or part of, Answers question referred by, Determines compliance with, Upholds all or part of, Overturns all or part of, andTransposes all or part of.

incoming_celex

String. The CELEX number of the incoming document (the document that the edge leads to).

incoming_node_type_id

Numeric. An ID number that uniquely identifies the type of the incoming node.

incoming_node_type

String. The type of the incoming node (the document that the edge leads to). Possible values include: Treaty articles, Directives, Regulations. Decisions, Court of Justice judgments, Advocate general opinions, General Court judgments, National implementing measures, and National court decisions.